+++ src/problem002.cpp
@@ -1,11 +1,11 @@
 #include "problem002.hpp"
 #include "fibonacci.hpp"
 
-unsigned long long problem002(const int n) {
-  auto fib{0ULL};
-  auto sum{0ULL};
+long problem002(const int n) {
+  auto fib{0L};
+  auto sum{0L};
 
-  for (auto i = 0; fib < static_cast<unsigned long long>(n); ++i) {
+  for (auto i = 0; fib < static_cast<long>(n); ++i) {
     fib = fibonacci(i);
     if (fib % 2 == 0)
       sum += fib;
