+++ src/problem002.hpp
@@ -1,3 +1,3 @@
 #pragma once
 
-unsigned long long problem002(const int n);
+long problem002(const int n);
